# spring/src/main/java/com/alvaro/demo/User/UserRestController.java
package com.alvaro.demo.User;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import ch.qos.logback.core.model.Model;

@RestController
@RequestMapping("/api")
public class UserRestController {
    
    @Autowired
    private UserService userService;

    @GetMapping("/users")
    public ResponseEntity<List<User>> users(Model model){
        List<User> listUsers = this.userService.getAll();
        if(listUsers.isEmpty()) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        else return new ResponseEntity<>(listUsers,HttpStatus.OK);
    }

    @GetMapping("/users/{id}")
    public ResponseEntity<User> usersId(Model model, @PathVariable long id){
        User user = this.userService.getById(id);
        if(user == null) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        else return new ResponseEntity<>(user,HttpStatus.OK);
    }

    @DeleteMapping("/users/{id}")
    public ResponseEntity<User> deleteUser(Model model, @PathVariable long id){
        User userDelete = userService.deleteUser(id);
        if(userDelete != null) return new ResponseEntity<>(userDelete,HttpStatus.OK);
        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
    }

}