# spring/src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deceye API",
    "version": "1.0.13"
  },
  "servers": [
    {
      "url": "http://localhost:6868/api",
      "description": "development"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Get All Users.",
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/user"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get Especific User.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the user to get"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      },
      "delete": {
        "summary": "Delete Especific User.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the user to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      }
    },
    "/api/entries": {
      "get": {
        "summary": "Get All Entries.",
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/entry"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      }
    },
    "/api/entries/{id}": {
      "get": {
        "summary": "Get Especific Entries.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the entry to get"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/entry"
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      },
      "delete": {
        "summary": "Delete Especific Entry.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the entry to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/entry"
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      }
    },
    "/api/comments": {
      "get": {
        "summary": "Get All Comments.",
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/comment"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      }
    },
    "/api/comments/{id}": {
      "get": {
        "summary": "Get Especific Comment.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the comment to get"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      },
      "delete": {
        "summary": "Delete Especific Comment.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the comment to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comment"
                }
              }
            }
          },
          "404": {
            "description": "Bad request."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "user": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "ROLE_USER|ROLE_ADMIN"
            }
          },
          "email": {
            "type": "string"
          }
        }
      },
      "entry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "fecha": {
            "type": "string",
            "example": "19-04-2023"
          },
          "titulo": {
            "type": "string"
          },
          "autor": {
            "type": "string"
          },
          "contenido": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/user"
          }
        }
      },
      "comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "fecha": {
            "type": "string",
            "example": "19-04-2023"
          },
          "autor": {
            "type": "string"
          },
          "contenido": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/user"
          }
        }
      }
    }
  }
}